Make the complete capture area obey the color scheme
By making the complete capture area follow the color of the scheme, the dark scheme now avoids drawing bright white areas. At least for my OS-global dark theme the readability is greatly improved, because it is not drawing grey text on white background even in the dark scheme any more. Light theme still works, of course, I have just added some contrast to it and changed it to be based on mPlotBackgroundColor.